+++ dotnet/src/webdriver/BiDi/Communication/Json/Converters/DateTimeOffsetConverter.cs
@@ -27,16 +27,7 @@ internal class DateTimeOffsetConverter : JsonConverter<DateTimeOffset>
 {
     public override DateTimeOffset Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
     {
-        // Workaround: it should be Int64, chrome uses double for `expiry` like "expiry":1737379944.308351
-
-        if (reader.TryGetInt64(out long unixTime) is false)
-        {
-            var doubleValue = reader.GetDouble();
-
-            unixTime = Convert.ToInt64(doubleValue);
-        }
-
-        return DateTimeOffset.FromUnixTimeMilliseconds(unixTime);
+        return DateTimeOffset.FromUnixTimeMilliseconds(reader.GetInt64());
     }
 
     public override void Write(Utf8JsonWriter writer, DateTimeOffset value, JsonSerializerOptions options)