+++ dotnet/src/webdriver/BiDi/Session/SubscribeCommand.cs
@@ -22,10 +22,10 @@ using System.Collections.Generic;
 
 namespace OpenQA.Selenium.BiDi.Session;
 
-internal sealed class SubscribeCommand(SubscribeCommandParameters @params)
-    : Command<SubscribeCommandParameters, SubscribeResult>(@params, "session.subscribe");
+internal sealed class SubscribeCommand(SubscribeParameters @params)
+    : Command<SubscribeParameters, SubscribeResult>(@params, "session.subscribe");
 
-internal sealed record SubscribeCommandParameters(IEnumerable<string> Events, IEnumerable<BrowsingContext.BrowsingContext>? Contexts) : CommandParameters;
+internal sealed record SubscribeParameters(IEnumerable<string> Events, IEnumerable<BrowsingContext.BrowsingContext>? Contexts) : Parameters;
 
 public sealed class SubscribeOptions : CommandOptions
 {